+++ stacActions.config.js
@@ -1,7 +1,7 @@
 // import StacMap from './src/actions/stac/StacMap.js';
-// import FavoriteCard from './src/actions/stac/FavoriteCard.js';
-
+import FavoriteCard from './src/actions/stac/FavoriteCard.js';
+FavoriteCard.showInBrowse = false; // Show only in search, not in browsing mode
 export default {
-  // FavoriteCard,
+  FavoriteCard,
   // StacMap,
 };